finsrc/requirements.txt
kafka-python==1.3.4
python-dateutil==2.6.1
pytz==2017.2
six==1.10.0
ystockquote==0.2.5
